Improve BKD

+++ src/core/analysis/token_attributes/char_term_attribute_impl.rs
@@ -64,7 +64,7 @@ impl CharTermAttributeImpl {
     pub fn char_at(&self, index: usize) -> Result<char> {
         debug_assert!(index <= i32::MAX as usize);
         debug_assert!(self.term_length <= i32::MAX as usize);
-        CoreHelper::check_index(index as i32, self.term_length as i32)?;
+        CoreHelper::check_index(index, self.term_length)?;
         Ok(self.term_buffer[index])
     }
     pub fn sub_sequence(&self, start: usize, end: usize) -> Result<&[char]> {
