+++ src/core/codecs/dummy/dummy_mutable_point_tree.rs
@@ -52,7 +52,7 @@ impl PointTree for DummyMutablePointTree {
         unreachable!("Dummy implementation: this method should never be called in real usage")
     }
 
-    fn size(&self) -> Result<i64> {
+    fn size(&self) -> Result<usize> {
         unreachable!("Dummy implementation: this method should never be called in real usage")
     }
