+++ src/core/document/float_field.rs
@@ -38,7 +38,7 @@ pub mod float_field_type {
 
     pub static FIELD_TYPE: Lazy<FieldType> = Lazy::new(|| {
         let mut ft = FieldType::new();
-        ft.set_dimensions(1, BitUtil::FLOAT_BYTES as i32)
+        ft.set_dimensions(1, BitUtil::FLOAT_BYTES)
             .expect("set_dimensions should not fail");
         ft.set_doc_values_type(DocValuesType::SortedNumeric)
             .expect("set_doc_values_type should not fail");
