+++ src/core/document/int_field.rs
@@ -36,7 +36,7 @@ pub mod int_field_type {
 
     pub static FIELD_TYPE: Lazy<FieldType> = Lazy::new(|| {
         let mut ft = FieldType::new();
-        ft.set_dimensions(1, BitUtil::INT_BYTES as i32)
+        ft.set_dimensions(1, BitUtil::INT_BYTES)
             .expect("set_dimensions should not fail");
         ft.set_doc_values_type(DocValuesType::SortedNumeric)
             .expect("set_doc_values_type should not fail");
