+++ src/core/index/tracking_tmp_output_directory_wrapper.rs
@@ -94,7 +94,7 @@ where
         self.base.delete_file(name)
     }
 
-    fn file_length(&self, name: &str) -> Result<i64> {
+    fn file_length(&self, name: &str) -> Result<usize> {
         self.base.file_length(name)
     }
