+++ src/core/store/filter_directory.rs
@@ -96,7 +96,7 @@ where
         self.delegate.delete_file(name)
     }
 
-    fn file_length(&self, name: &str) -> Result<i64> {
+    fn file_length(&self, name: &str) -> Result<usize> {
         self.delegate.file_length(name)
     }
