+++ src/core/util/numeric_utils.rs
@@ -99,14 +99,14 @@ impl NumericUtils {
 
     /// Result = a - b, where a >= b, else `LuceneError` is returned.
     pub fn subtract(
-        bytes_per_dim: i32,
-        dim: i32,
+        bytes_per_dim: usize,
+        dim: usize,
         a: &[u8],
         b: &[u8],
         result: &mut [u8],
     ) -> Result<()> {
-        let start = (dim * bytes_per_dim) as usize;
-        let end = start + bytes_per_dim as usize;
+        let start = dim * bytes_per_dim;
+        let end = start + bytes_per_dim;
         let mut borrow = 0;
 
         for i in (start..end).rev() {
@@ -821,7 +821,7 @@ mod tests {
 
             let mut result = vec![0u8; num_bytes];
 
-            NumericUtils::subtract(num_bytes as i32, 0, &v1_bytes, &v2_bytes, &mut result)?;
+            NumericUtils::subtract(num_bytes, 0, &v1_bytes, &v2_bytes, &mut result)?;
 
             let v1_clone = v1.clone();
             let v2_clone = v2.clone();
